Crate nexus_sdk

An SDK.

Modules

compile

Configure the dynamic compilation of guest programs.

error

Contains error types for SDK-specific interfaces.

hypernova

Interface into proving with HyperNova.

jolt

Experimental interface into proving with Jolt.

nova

Interface into proving with Nova.

Enums

Local

Indicator type that local compute will be used for proving the zkVM.

Traits

Compute

A compute resource.

Parameters

A parameter set used for proving and verifying.

Prover

A prover (and runner) for the zkVM.

Verifiable

A verifiable proof of a zkVM execution. Also contains a view capturing the output of the machine.