Trait nexus_sdk::Parameters

pub trait Parameters {
    type Error;

    // Required methods
    fn generate_for_testing() -> Result<Self, Self::Error>
       where Self: Sized;
    fn load(path: &Path) -> Result<Self, Self::Error>
       where Self: Sized;
    fn save(pp: &Self, path: &Path) -> Result<(), Self::Error>;
}

A parameter set used for proving and verifying.

Required Associated Types

type Error

Required Methods

fn generate_for_testing() -> Result<Self, Self::Error>where Self: Sized,

Generate testing parameters.

In deployment, prover parameters often depend on an external reference, like a structured reference string (SRS). As such, individual provers may expose an interface for generating production parameters (e.g., HyperNova::Generate).

fn load(path: &Path) -> Result<Self, Self::Error>where Self: Sized,

Load parameters from a file.

fn save(pp: &Self, path: &Path) -> Result<(), Self::Error>

Save parameters to a file.

Object Safety

This trait is not object safe.