Trait nexus_sdk::Prover

pub trait Prover {
    type Memory;
    type Params: Parameters;
    type Error;

    // Required methods
    fn new(elf_bytes: &[u8]) -> Result<Self, Self::Error>
       where Self: Sized;
    fn compile(opts: &CompileOpts) -> Result<Self, Self::Error>
       where Self: Sized,
             Self::Error: From<Error>;
    fn run<T, U>(self, input: Option<T>) -> Result<View<U>, Self::Error>
       where T: Serialize + Sized,
             U: DeserializeOwned;
    fn prove<T, U>(
        self,
        pp: &Self::Params,
        input: Option<T>
    ) -> Result<impl Verifiable, Self::Error>
       where T: Serialize + Sized,
             U: DeserializeOwned;

    // Provided method
    fn new_from_file(path: &PathBuf) -> Result<Self, Self::Error>
       where Self: Sized,
             Self::Error: From<Error> { ... }
}

A prover (and runner) for the zkVM.

Required Associated Types

type Memory

type Params: Parameters

type Error

Required Methods

fn new(elf_bytes: &[u8]) -> Result<Self, Self::Error>where Self: Sized,

Construct a new proving instance from raw ELF bytes.

fn compile(opts: &CompileOpts) -> Result<Self, Self::Error>where Self: Sized, Self::Error: From<Error>,

Construct a new proving instance through dynamic compilation (see compile).

fn run<T, U>(self, input: Option<T>) -> Result<View<U>, Self::Error>where T: Serialize + Sized, U: DeserializeOwned,

Run the zkVM on input of type T and return a view of the execution output by deserializing the output tape as of type U.

fn prove<T, U>( self, pp: &Self::Params, input: Option<T> ) -> Result<impl Verifiable, Self::Error>where T: Serialize + Sized, U: DeserializeOwned,

Prove the zkVM on input of type T and return a verifiable proof, along with a view of the execution output by deserializing the output tape as of type U.

Provided Methods

fn new_from_file(path: &PathBuf) -> Result<Self, Self::Error>where Self: Sized, Self::Error: From<Error>,

Construct a new proving instance by reading an ELF file.

Object Safety

This trait is not object safe.